// src/services/api/generated/movie-service/model/createMovieRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Cinema Booking System API
 * API documentation for Cinema Booking System
 * OpenAPI spec version: 1.0.0
 */
import type { CreateMovieRequestStatus } from './createMovieRequestStatus';

export interface CreateMovieRequest {
  /**
     * @minLength 0
     * @maxLength 255
     */
  title: string;
  /**
     * @minLength 0
     * @maxLength 5000
     */
  description?: string;
  /** @minimum 1 */
  durationMinutes: number;
  releaseDate?: string;
  /**
     * @minLength 0
     * @maxLength 500
     */
  posterUrl?: string;
  /**
     * @minLength 0
     * @maxLength 500
     */
  trailerUrl?: string;
  status: CreateMovieRequestStatus;
  /** @minItems 1 */
  genreIds: string[];
}